Move integration functions to own module and add 'add_slots'

+++ src/srv/src/data/user.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 
 /// Code uniquely identifying a user
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct UserId(pub u32);
+pub struct UserId(pub u64);
 
 impl std::fmt::Display for UserId {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
